Validator extenstion expanded, newUserDTO created
Validator.password written for validating any type of password, Validatble is an abstract class which provides validate() method which check validation and throw an error when something is invalid. validateFunc function is provided for classes which can't extends any other class. Tests are needed to Validator. NewUserDTO is written, by not tested yet.