A lot happened here
I've forgotten to commit so I committing everything at once. Validation was rebuild - now it is a ValidationService not decentralized validate() methods. Error hierarchy created. PersistedUserDTO created. UserDTOMapper. UserService created.  All is tested. There is also fragment of UserRESTAPI, but this is messy and not tested (was created for presentation purpose only) and it will be rewrite soon.